chore: remove tslint
The fp-ts version bump triggered numerous

```
ERROR: eqString is deprecated
```

deprecation workings. Tslint is very old an deprecated and this is
enough of a pain to convince me to rip out tslint and we can replace
that with eslint if/when this project gets enough additional TLC to
warrant spending time on a linter.